requirements.txt
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
SQLAlchemy==1.4.47
Werkzeug==2.2.3
python-dotenv==0.19.2
pytest==7.1.2
pytest-flask==1.2.0
aiohttp==3.8.1